backend/pyproject.toml
[tool.poetry]
name = "divez"
version = "0.1.0"
description = ""
authors = ["Guilherme <gui.santos.sa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.30.0"
alembic = "^1.13.0"
pydantic = {extras = ["email"], version = "^2.11.7"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.10"
pydantic-settings = "^2.10.1"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
